gitflow-feature-stash: ipynb_workflow
Changes are intended to support the cleanup of old MetaAnalysis stubs that are no longer needed following the fork, and development of a single clean branch for running the analysis and documentation of the workflow.

First pass is removing unneeded documents associated with the NER workflow and dockerization of the workflow (we'll add docker back in later if needed).