Renamed and cleaned `data_preprocessing`.
The filename and function were changed to use the underscore. There was a lot of flagged things to check previously. I cleaned it all up to meet our current use-case. Added BeautifulSoup parser to remove HTML like tags.
The function still feels a bit too "tailored" to the dataset itself but works okay.