Added a function to push labels to S3.
The function uses `source` and a tagging system. We could add other functions to help us manage the way labels are added, but generally we've got a function that can add a set of labels up to S3 in a parquet file. We do some duplicate cleaning based on DOI, source and label, but otherwise it just accepts the new set of labels.